+++ tests/modules/test_dielectricsphere.py
@@ -13,7 +13,9 @@ from pathlib import Path
 import MDAnalysis as mda
 import numpy as np
 import pytest
+import sympy as sp
 from numpy.testing import assert_allclose
+from util import error_prop
 
 from maicos import DielectricSphere
 
@@ -22,6 +24,7 @@ sys.path.append(str(Path(__file__).parents[1]))
 from data import (  # noqa: E402
     DIPOLE_GRO,
     DIPOLE_ITP,
+    WATER_2F_TRR_NPT,
     WATER_GRO_NPT,
     WATER_TPR_NPT,
     WATER_TRR_NPT,
@@ -64,6 +67,12 @@ class TestDielectricSphere(object):
     def ag_single_frame(self):
         """Import MDA universe, single frame."""
         u = mda.Universe(WATER_TPR_NPT, WATER_GRO_NPT)
+        return u.atoms
+
+    @pytest.fixture()
+    def ag_two_frames(self):
+        """Import MDA universe, single frame."""
+        u = mda.Universe(WATER_TPR_NPT, WATER_2F_TRR_NPT)
         return u.atoms
 
     @pytest.mark.parametrize("selection", (1, 2))
@@ -167,3 +176,30 @@ class TestDielectricSphere(object):
         ag = mda.Universe(DIPOLE_ITP, DIPOLE_GRO, topology_format="itp").atoms
         DielectricSphere(ag, rmin=rmin, rmax=rmax)
         assert result == (warning in "".join([rec.message for rec in caplog.records]))
+
+    def test_error_calculation(self, ag_two_frames):
+        """Test the analytic error propagation"""
+        eps = DielectricSphere(ag_two_frames).run()
+        eps.means.mM_r = 0.6
+        eps.sems.mM_r = 0.2
+
+        eps.means.m_r = 0.4
+        eps.sems.m_r = 0.1
+
+        eps.means.M_r = 0.2
+        eps.sems.M_r = 0.01
+
+        # rerun conclude function with changed values
+        eps._conclude()
+        mM_r, m_r, M_r = sp.symbols("mM_r m_r M_r")
+
+        cov_rad_sympy = mM_r - m_r * M_r
+
+        eps_rad_sympy = 1 - (
+            4 * np.pi * eps.results.bin_pos[0] ** 2 * eps.pref * cov_rad_sympy
+        )
+
+        deps_rad_sympy = error_prop(
+            eps_rad_sympy, [mM_r, m_r, M_r], [eps.sems.mM_r, eps.sems.m_r, eps.sems.M_r]
+        )(eps.means.mM_r, eps.means.m_r, eps.means.M_r)
+        assert_allclose(eps.results.deps_rad[0], deps_rad_sympy)