+++ CHANGELOG.rst
@@ -14,8 +14,9 @@ The rules for MAICoS' CHANGELOG file:
 v0.9 (XXXX/XX/XX)
 -----------------
 Philip Loche, Marc Sauter, Kira Fischer, Federico Grasselli, Henrtik Stooß,
-Adyant Agrawal
+Adyant Agrawal
 
+- Added tests of the analytical error propagation (!292)
 - Raise an error if ``pack`` is ``True`` and ``refgroup`` is not ``None`` (!311)
 - Use glob for find modules in docs (!312)
 - Distributing tests across multiple CPUs (!309)